【PR】: sync ge_dev to master 20260325

Created-by: peiyang
Commit-by: wtf_newage;ling-DT;yelongjian;isaacxu;Mexyy
Merged-by: peiyang
Description: 

# Pull Request

## 描述
本PR指在将develop分支的代码同步合入master分支，包括以下特性：
1、llt任务打印详细日志，方便出现失败时能和成功日志对比
2、根据rts接口判断是否需要拆流
3、修复graph_name带非法字符场景导致dump报错

## 变更类型
请选择本次引入的变更类型：
<!--  [x] 表示选中 -->
- [ ] 🐛 Bug 修复
- [ ] ✨ 新功能
- [ ] 💄 代码风格更新（格式化，局部变量）
- [ ] ♻️ 重构（既不修复错误也不增加功能的代码变动）
- [ ] 📦 构建过程或辅助工具的变动
- [ ] 📝 文档内容更新

## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在当前页面的右侧'关联Issue'部分添加相应Issue链接，并勾选'合并后关闭已关联的 Issue'选项。 -->


## 如何测试
描述测试此变更的步骤和前提条件：
1.
2.

## 核对清单
<!--  [x] 表示选中 -->
- [ ] 我的代码遵循了项目的代码风格
- [ ] 我已对代码进行了自测
- [ ] 我已更新了相关的文档
- [ ] 我在标题中使用了合适的类型标签（如：`feat:`, `fix:`）
- [ ] 我已经详细阅读了贡献指南（CONTRIBUTING.md），并遵守了其中的所有规定，包括但不限于commit message的格式、无效commit的合并等

## 其他信息
在此添加任何其他关于本次 PR 的说明。




See merge request: cann/ge!1475

+++ compiler/graph/optimize/autofuse/autofuse/lowering/lowering_utils.h
@@ -21,7 +21,7 @@
 
 namespace ge {
 struct LoweringConfig {
-  size_t max_loop_ops = 64U;
+  uint64_t max_loop_ops = 64U;
   size_t max_loop_loads = 4U;
   size_t max_buffer_readers = 4U;
 };
